chore: Added tentant to OAuth token storage

+++ App.Lib.Database/src/OAuthTokenStorage.cs
@@ -13,12 +13,12 @@ public class OAuthTokenStorage : IOAuthTokenStorage
         _dbContext = dbContext;
     }
 
-    public async Task<IOAuthToken> Get(string name)
+    public async Task<IOAuthToken> Get(string name, Guid tenant)
     {
         var tokenCollection = _dbContext.GetCollection<OAuthTokenDocument>();
         var token = await tokenCollection
             .AsQueryable()
-            .Where(d => d.Name == name)
+            .Where(d => d.Name == name && d.Tenant == tenant)
             .FirstOrDefaultAsync();
 
         if (token != null)
@@ -27,10 +27,12 @@ public class OAuthTokenStorage : IOAuthTokenStorage
         }
 
         // Insert new token but be sure not to override any already existing
-        token = new OAuthTokenDocument(name);
+        token = new OAuthTokenDocument(name, tenant);
         await tokenCollection.UpdateOneAsync(
-            filter => filter.Name == name,
-            Builders<OAuthTokenDocument>.Update.Set(field => field.Name, name),
+            filter => filter.Name == name && filter.Tenant == tenant,
+            Builders<OAuthTokenDocument>.Update
+                .Set(field => field.Name, name)
+                .Set(field => field.Tenant, tenant),
             new UpdateOptions { IsUpsert = true }
         );
 
@@ -41,14 +43,15 @@ public class OAuthTokenStorage : IOAuthTokenStorage
     {
         if (string.IsNullOrEmpty(token.AccessToken))
         {
-            await Delete(token.Name);
+            await Delete(token.Name, token.Tenant);
             return;
         }
 
         await _dbContext.GetCollection<OAuthTokenDocument>().UpdateOneAsync(
-            filter => filter.Name == token.Name,
+            filter => filter.Name == token.Name && filter.Tenant == token.Tenant,
             Builders<OAuthTokenDocument>.Update
                 .Set(field => field.Name, token.Name)
+                .Set(field => field.Tenant, token.Tenant)
                 .Set(field => field.AccessToken, token.AccessToken)
                 .Set(field => field.RefreshToken, token.RefreshToken)
                 .Set(field => field.ExpiresAt, token.ExpiresAt),
@@ -56,9 +59,9 @@ public class OAuthTokenStorage : IOAuthTokenStorage
         );
     }
 
-    public async Task Delete(string tokenName)
+    public async Task Delete(string tokenName, Guid tenant)
     {
         await _dbContext.GetCollection<OAuthTokenDocument>()
-            .DeleteOneAsync(filter => filter.Name == tokenName);
+            .DeleteOneAsync(filter => filter.Name == tokenName && filter.Tenant == tenant);
     }
 }
